% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_pop_weighted_score}
\alias{aggregate_pop_weighted_score}
\title{Population-weighted scores}
\usage{
aggregate_pop_weighted_score(data, x, higher_id, lower_pop)
}
\arguments{
\item{data}{A dataframe containing a variable to be aggregated, lower level
geography population estimates, and a higher level geographical grouping
variable.}

\item{x}{Name of the variable in the dataframe containing the variable to be
aggregated (e.g., score) for the lower level geography.}

\item{higher_id}{Name of the variable in the dataframe containing the higher
level geography names/codes.}

\item{lower_pop}{Name of the variable in the dataframe containing the
population estimates of the lower level geography.}
}
\description{
Aggregate population-weighted scores within small areas.
}
\examples{
\dontrun{
aggregate_pop_weighted_score(
  example_aggregate_data,
  score,
  higher_geography_code,
  lower_geography_population
)
}
}
