% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_proportion}
\alias{aggregate_proportion}
\title{Calculate the 'proportion' scores when aggregating up small areas}
\usage{
aggregate_proportion(data, x, higher_id)
}
\arguments{
\item{data}{A dataframe containing a variable to be aggregated and a higher
level geographical grouping variable.}

\item{x}{Name of the variable in the dataframe for which you want to
calculate proportions. It must have only two possible values.}

\item{higher_id}{Name of the variable in the dataframe containing the higher
level geography names/codes.}
}
\description{
Calculate proportion of small areas in the higher-level geography that are
within the 10\% most deprived areas in the nation.
}
\examples{
\dontrun{
aggregate_proportion()
}
}
