% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_austen_data}
\alias{get_austen_data}
\title{Example data from \pkg{janeaustenr}}
\usage{
get_austen_data(x)
}
\description{
Example data from \pkg{janeaustenr}
}
\examples{
\dontrun{
austen_books <- get_sample_data()
}
}
