% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sample_data}
\alias{get_sample_data}
\title{Sample data from \pkg{janeaustenr}}
\usage{
get_sample_data(x)
}
\description{
Sample data from \pkg{janeaustenr}
}
\examples{
\dontrun{
austen_books <- get_sample_data()
}
}
