% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranks.R
\name{inverse_rank}
\alias{inverse_rank}
\title{Inverse rank}
\usage{
inverse_rank(vec)
}
\arguments{
\item{vec}{Vector to be inverse ranked.}
}
\description{
\code{inverse_rank} returns the inverse rank of a vector (i.e., the
highest value receives a rank of 1).
}
\examples{
inverse_rank(c(1:3))
}
