% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert_rank}
\alias{invert_rank}
\title{Invert rank}
\usage{
invert_rank(x)
}
\arguments{
\item{x}{Vector to be inverse ranked.}
}
\description{
\code{invert_rank} returns the inverse rank of a vector (i.e., the
highest value receives a rank of 1).
}
\examples{
invert_rank(c(1:3))
}
