% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert_this}
\alias{invert_this}
\title{Invert this}
\usage{
invert_this(x)
}
\arguments{
\item{x}{Vector to be inverted.}
}
\description{
\code{invert_this} returns the inverse of a vector. For example, a decile of
10 becomes 1, and a decile 1 becomes 10. NA's are removed by default.
}
\examples{
invert_this(c(1:3))
}
