% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_na}
\alias{keep_na}
\title{Keep NA}
\usage{
keep_na(data)
}
\arguments{
\item{data}{Data frame}
}
\description{
\code{keep_na} returns all rows where any column contains a missing value
}
\examples{
\dontrun{
keep_na(mtcars)
}
}
