% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytextprep.R
\name{lazytextprep}
\alias{lazytextprep}
\title{clean and/or tokenize your text data in a single function}
\usage{
lazytextprep(
  textdata,
  textvar,
  type = "docs",
  language = "english",
  outdir = NA,
  outname = "/transformations.txt"
)
}
\description{
Takes a text variable from a dataframe and runs a number of standard text preprocessing procedures on it, like removing html tags, removing stopwords, converting to lowercase. Preprocessing techniques and tokenization are applied in an interactive yes/no console session with the user. A list of the procedures used are saved in a local .txt file in directory specified by the user.
}
\details{
@param textdata a dataframe containing a text variable
@param textvar the name of the column in the first param containing text
@param type right now there is only one type called "docs"
@param language user specified language, determines what tm::stopword dictionary is used
@param outdir the directory that the user wishes to have the output .txt file saved in
@param outname the name of the transformations .txt summary file, defaults to transformations.txt but can be renamed

@return the dataframe with a cleaned and/or tokenized text variable

@export

@import textclean
@import dplyr
@import tidytext
@import textstem
@import SnowballC
@import tibble
@import tm
}
