% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{normalise_mean}
\alias{normalise_mean}
\title{Mean normalisation}
\usage{
normalise_mean(x, remove_na = TRUE)
}
\arguments{
\item{x}{Vector to standardise.}

\item{remove_na}{Boolean indicating whether NA values should be removed.
Defaults to TRUE.}
}
\description{
\code{normalise_mean} standardises (i.e., normalises) a vector to mean = 0 &
SD = 1.
}
\examples{
normalise_mean(c(1:10))
}
