% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{normalise_median}
\alias{normalise_median}
\title{Median normalisation}
\usage{
normalise_median(x, remove_na = TRUE)
}
\arguments{
\item{x}{A numeric vector that will be normalised.}

\item{remove_na}{Boolean indicating whether NA values should be removed.
Defaults to TRUE.}
}
\value{
A numeric vector of the same length as \code{x} where each element has
been positionally normalised.
}
\description{
This function performs median normalisation on a numeric vector. It
subtracts the median from each element in the vector and then divides the
resulting values by the square root of the sum of their squared deviations
from the median.
}
\examples{
normalise_median(c(1, 2, 3, 4, 5))
normalise_median(c(10, 20, 30, 40, 50))

}
