% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise_rank}
\alias{normalise_rank}
\title{Normalise ranks}
\usage{
normalise_rank(vec)
}
\arguments{
\item{vec}{Vector of ranks to be normalised.}
}
\description{
\code{normalise_rank} normalises a rank between a scale of 0 and 1.
}
\examples{
normalise_rank(c(1:3))
}
