% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise_ranks_01}
\alias{normalise_ranks_01}
\title{Normalise ranks between 0 and 1}
\usage{
normalise_ranks_01(x)
}
\arguments{
\item{x}{Vector of ranks to be scaled.}
}
\description{
\code{normalise_ranks_01} scales (i.e., normalises) a rank between a scale of
0 and 1.
}
\examples{
normalise_ranks_01(c(1:3))
}
