% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{positional_normalisation}
\alias{positional_normalisation}
\title{Positional Normalisation}
\usage{
positional_normalisation(x)
}
\arguments{
\item{x}{A numeric vector that will be normalised.}
}
\value{
A numeric vector of the same length as \code{x} where each element has been
positionally normalised.
}
\description{
This function performs positional normalisation on a numeric vector.
It subtracts the median from each element in the vector and then divides
the resulting values by the square root of the sum of their squared deviations from the median.
}
\examples{
positional_normalisation(c(1, 2, 3, 4, 5))
positional_normalisation(c(10, 20, 30, 40, 50))

}
