% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textprep_tree.R
\name{textprep_tree}
\alias{textprep_tree}
\title{Get decision tree for textprep function}
\usage{
textprep_tree(x)
}
\description{
Generates decision tree for textprep function
}
\details{
@return a decision tree for the function textprep

@export

@import DiagrammeR
}
