% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_exp}
\alias{transform_exp}
\title{Exponentially transform domain indexes}
\usage{
transform_exp(x)
}
\arguments{
\item{x}{A vector of integers, in the range \[0, 1\]}
}
\description{
Transform domains to an exponential distribution using the exponential
transformation function listed in Appendix A of the Welsh IMD's technical
report: \url{https://www.gov.wales/sites/default/files/statistics-and-research/2023-10/welsh-index-multiple-deprivation-2019-technical-report.pdf}.
}
\section{When to use}{

The exponential transformation of ranks makes it harder for high deprivation
in one domain to be offset by low deprivation in another. This transformation
stretches out the rankings of the most deprived areas, creating gaps that
highlight the true differences and place more emphasis on the most deprived
parts of the distribution.
}

\examples{
transform_exp(c(0, 0.5, 1))
}
