% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantise.R
\name{quantise}
\alias{quantise}
\title{Quantise}
\usage{
quantise(x, num_quantiles = 10, invert = FALSE)
}
\arguments{
\item{x}{The vector to quantise.}

\item{num_quantiles}{The Number of quantiles, defaults to 10.}

\item{invert}{Should the highest quantile represent the lowest input value?
Defaults to false.}
}
\description{
\code{quantise} quantises a vector.
}
\examples{
quantise(c(1:20))
quantise(c(1:20), num_quantiles = 10, invert = TRUE)
}
