% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{standardise}
\alias{standardise}
\title{Standardise}
\usage{
standardise(vec)
}
\arguments{
\item{vec}{Vector to normalise}
}
\description{
\code{standardise} standardises (i.e., normalises) a vector to mean = 0 & SD
= 1.
}
\examples{
standardise(c(1:10))
}
